+++ apps/oxfmt/Cargo.toml
@@ -28,7 +28,6 @@ doctest = false
 
 [dependencies]
 oxc_allocator = { workspace = true, features = ["pool"] }
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
 oxc_config = { workspace = true }
 oxc_diagnostics = { workspace = true }
 oxc_formatter = { workspace = true }
@@ -92,7 +91,3 @@ default = ["napi"]
 napi = ["dep:napi", "dep:napi-derive"]
 allocator = ["dep:mimalloc-safe"]
 detect_code_removal = ["oxc_formatter/detect_code_removal"]
-
-[package.metadata.cargo-shear]
-# `oxc_ast` is a feature-only dependency, to enable its `disable_old_builder` feature
-ignored = ["oxc_ast"]
