+++ crates/oxc/Cargo.toml
@@ -43,12 +43,6 @@ oxc_syntax = { workspace = true }
 oxc_transformer = { workspace = true, optional = true }
 oxc_transformer_plugins = { workspace = true, optional = true }
 
-[dev-dependencies]
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
-
 [features]
 default = ["regular_expression"]
