+++ crates/oxc_ast/src/builder/mod.rs
@@ -1,30 +1,16 @@
 //! AST builder.
 //!
-//! This is undergoing change at present.
-//!
-//! Explanation of the motivation for this change here: <https://github.com/oxc-project/oxc/issues/23043>.
-//!
-//! ## Old builder
-//!
-//! [`AstBuilder`] used to be a [`Copy`] type, which had own methods for:
-//!
-//! * Creating AST nodes e.g. `builder.statement_expression(span, expr)`
-//! * Creating primitives e.g. `builder.vec()`, `builder.ident(str)`
-//!
-//! These methods are defined in `../generated/ast_builder.rs` and `methods.rs`.
-//!
-//! ## New builder
-//!
-//! We have now added methods to AST types themselves which perform the same role,
-//! and are passed an `&B where B: GetAstBuilder` or `&A where A: GetAllocator`:
+//! AST nodes are created by builder methods defined on the AST types themselves,
+//! which are passed a `&B where B: GetAstBuilder` or `&A where A: GetAllocator`:
 //!
 //! * `Statement::new_expression_statement(span, expr, &builder)`
 //! * `Vec::new_in(&builder)`, `Ident::from_str_in(str, &builder)`
 //!
-//! `AstBuilder` is no longer `Copy` or `Clone`, and is passed by reference.
-//! Its `allocator` field is no longer public. Use `allocator` method provided by `GetAllocator` trait.
+//! [`AstBuilder`] provides the memory arena and [`NodeId`]s that these methods use.
+//! It is not [`Copy`] or [`Clone`], and is passed by reference.
+//! Its `allocator` field is private - use the `allocator` method provided by the [`GetAllocator`] trait.
 //!
-//! Implementing `GetAstBuilder` on types which hold an `AstBuilder` allows for a shorter syntax:
+//! Implementing [`GetAstBuilder`] on types which hold an [`AstBuilder`] allows for a shorter syntax:
 //!
 //! * Long: `Vec::new_in(&self.ast)`
 //! * Short: `Vec::new_in(self)`
@@ -78,30 +64,21 @@
 //!
 //! These AST type builder methods are defined in `../generated/builder_methods.rs` and `custom.rs`.
 //!
-//! ## Migration
+//! ## Migration from the old builder
 //!
-//! To minimize immediate breaking changes for downstream consumers, and allow them to migrate incrementally,
-//! at present `AstBuilder` still has its own methods, but can also be passed to the AST type builder methods.
+//! [`AstBuilder`] used to be a [`Copy`] type with its own methods for creating AST nodes
+//! (e.g. `builder.statement_expression(span, expr)`) and primitives (e.g. `builder.vec()`,`builder.ident(str)`).
+//! Those methods have now been removed. Use the AST type builder methods described above instead.
 //!
-//! Once a project has migrated, they should enable the `disable_old_builder` Cargo feature,
-//! which will remove the old-style own methods on `AstBuilder`.
+//! [`AstBuilder`] and [`NONE`] are no longer re-exported from the crate root either -
+//! import them from this module instead.
 //!
-//! After a few weeks, we will remove `AstBuilder`'s own methods entirely.
-//!
-//! ## Oxc
-//!
-//! All Oxc crates have been migrated to the new usage.
-//!
-//! `disable_old_builder` Cargo feature is enabled in all Oxc crates which utilize `AstBuilder`.
-//! Where those crates expose the `AstBuilder` they use to user code, the feature is only enabled in tests.
+//! Explanation of the motivation for this change here: <https://github.com/oxc-project/oxc/issues/23043>.
 
 use oxc_allocator::{Allocator, ArenaBox, ArenaVec, FromIn, GetAllocator};
 use oxc_syntax::node::NodeId;
 
 mod custom;
-
-#[cfg(not(feature = "disable_old_builder"))]
-mod methods;
 
 /// Trait for types which can create AST nodes.
 ///
@@ -138,15 +115,9 @@ pub trait GetAstBuilder<'a> {
 ///
 /// For use where no `NodeId`s are required on AST nodes as they are built
 /// e.g. parser, because `NodeId`s are assigned later when building `Semantic`.
-#[cfg_attr(not(feature = "disable_old_builder"), derive(Clone, Copy))]
 pub struct AstBuilder<'a> {
     /// The memory allocator used to allocate AST nodes in the arena.
-    #[cfg(feature = "disable_old_builder")]
     allocator: &'a Allocator,
-
-    /// The memory allocator used to allocate AST nodes in the arena.
-    #[cfg(not(feature = "disable_old_builder"))]
-    pub allocator: &'a Allocator,
 }
 
 impl<'a> AstBuilder<'a> {
