+++ crates/oxc_ast/src/lib.rs
@@ -56,8 +56,6 @@ mod generated {
 
     #[cfg(debug_assertions)]
     mod assert_layouts;
-    #[cfg(not(feature = "disable_old_builder"))]
-    mod ast_builder;
     mod builder_methods;
     mod derive_clone_in;
     mod derive_content_eq;
@@ -78,8 +76,6 @@ pub use generated::ast_kind;
 pub use ast::comment::{Comment, CommentContent, CommentKind, CommentPosition};
 pub use ast_kind::{AstKind, AstType};
 pub use ast_kind_impl::{MemberExpressionKind, ModuleDeclarationKind};
-#[cfg(not(feature = "disable_old_builder"))]
-pub use builder::{AstBuilder, NONE};
 pub use trivia::{
     CommentsRange, comments_range, get_comment_at, has_comments_between, is_inside_comment,
 };
