+++ crates/oxc_codegen/Cargo.toml
@@ -37,11 +37,6 @@ itoa = { workspace = true }
 rustc-hash = { workspace = true }
 
 [dev-dependencies]
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
-
 insta = { workspace = true }
 oxc_parser = { workspace = true }
 pico-args = { workspace = true }
