+++ crates/oxc_estree_tokens/Cargo.toml
@@ -28,9 +28,3 @@ oxc_estree = { workspace = true, features = ["serialize"] }
 oxc_parser = { workspace = true, features = ["mutate_tokens"] }
 oxc_span = { workspace = true, features = ["serialize"] }
 itoa = { workspace = true }
-
-[dev-dependencies]
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
