+++ crates/oxc_formatter/Cargo.toml
@@ -47,11 +47,6 @@ pico-args = { workspace = true }
 serde_json = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
-
 [build-dependencies]
 oxc_formatter_core = { workspace = true, features = ["test_harness"] }
