+++ crates/oxc_formatter_json/Cargo.toml
@@ -32,11 +32,6 @@ oxc_formatter_core = { workspace = true, features = ["test_harness"] }
 oxc_tasks_common = { workspace = true }
 pico-args = { workspace = true }
 
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
-
 [build-dependencies]
 oxc_formatter_core = { workspace = true, features = ["test_harness"] }
