+++ crates/oxc_jsdoc/Cargo.toml
@@ -24,10 +24,3 @@ oxc_ast = { workspace = true }
 oxc_span = { workspace = true }
 
 rustc-hash = { workspace = true }
-
-[dev-dependencies]
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
-
