+++ crates/oxc_linter/Cargo.toml
@@ -82,7 +82,3 @@ url = { workspace = true }
 insta = { workspace = true }
 markdown = { workspace = true }
 
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
