+++ crates/oxc_mangler/Cargo.toml
@@ -36,7 +36,3 @@ rustc-hash = { workspace = true }
 [dev-dependencies]
 oxc_parser = { workspace = true }
 
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
