+++ crates/oxc_minifier/Cargo.toml
@@ -44,11 +44,6 @@ oxc_codegen = { workspace = true, features = ["sourcemap"] }
 oxc_parser = { workspace = true }
 oxc_sourcemap = { workspace = true }
 
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
-
 insta = { workspace = true }
 javascript-globals = { workspace = true }
 pico-args = { workspace = true }
