+++ crates/oxc_parser/Cargo.toml
@@ -43,11 +43,7 @@ seq-macro = { workspace = true }
 memchr = { workspace = true }
 
 [dev-dependencies]
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder", "serialize"] }
-
+oxc_ast = { workspace = true, features = ["serialize"] }
 oxc_ast_visit = { workspace = true, features = ["serialize"] }
 pico-args = { workspace = true }
