+++ crates/oxc_semantic/Cargo.toml
@@ -45,11 +45,6 @@ oxc_parser = { workspace = true }
 rustc-hash = { workspace = true }
 serde_json = { workspace = true }
 
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
-
 [features]
 default = []
 cfg = ["dep:oxc_cfg"]
