+++ crates/oxc_transformer/Cargo.toml
@@ -58,11 +58,6 @@ oxc_codegen = { workspace = true, features = ["sourcemap"] }
 oxc_parser = { workspace = true }
 pico-args = { workspace = true }
 
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
-
 [features]
 # Run the experimental React Compiler as the first transform pass.
 react_compiler = ["dep:oxc_react_compiler"]
