+++ crates/oxc_transformer_plugins/Cargo.toml
@@ -49,7 +49,3 @@ oxc_sourcemap = { workspace = true }
 oxc_tasks_common = { workspace = true }
 oxc_transformer = { workspace = true }
 
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
