+++ crates/oxc_traverse/Cargo.toml
@@ -35,9 +35,3 @@ oxc_syntax = { workspace = true }
 
 itoa = { workspace = true }
 rustc-hash = { workspace = true }
-
-[dev-dependencies]
-# Catch usage of old `AstBuilder` APIs in tests, without affecting downstream consumers.
-# If we enabled `disable_old_builder` feature in main dependency, feature unification would
-# cause it to be activated for `oxc_ast` for the user too.
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
