+++ tasks/ast_tools/src/generators/mod.rs
@@ -5,7 +5,6 @@ use crate::{
 };
 
 mod assert_layouts;
-mod ast_builder;
 mod ast_kind;
 mod builder_methods;
 #[cfg(feature = "generate-js")]
@@ -29,7 +28,6 @@ mod visit;
 mod visit_js;
 
 pub use assert_layouts::AssertLayoutsGenerator;
-pub use ast_builder::AstBuilderGenerator;
 pub use ast_kind::AstKindGenerator;
 pub use builder_methods::BuilderMethodsGenerator;
 #[cfg(feature = "generate-js")]
