+++ tasks/benchmark/Cargo.toml
@@ -70,7 +70,7 @@ bench = false
 [dependencies]
 oxc = { workspace = true, optional = true, features = ["full"] }
 oxc_allocator = { workspace = true, optional = true }
-oxc_ast = { workspace = true, optional = true, features = ["disable_old_builder", "serialize"] }
+oxc_ast = { workspace = true, optional = true, features = ["serialize"] }
 oxc_ast_visit = { workspace = true, optional = true, features = ["serialize"] }
 oxc_codegen = { workspace = true, optional = true, features = ["sourcemap"] }
 oxc_estree_tokens = { workspace = true, optional = true }
