+++ tasks/rulegen/Cargo.toml
@@ -15,7 +15,7 @@ doctest = false
 
 [dependencies]
 oxc_allocator = { workspace = true }
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
+oxc_ast = { workspace = true }
 oxc_ast_visit = { workspace = true }
 oxc_parser = { workspace = true }
 oxc_span = { workspace = true }
@@ -32,7 +32,6 @@ proc-macro2 = { workspace = true, features = ["span-locations"] }
 
 [dev-dependencies]
 insta = { workspace = true }
-
 
 [package.metadata.cargo-shear]
 ignored = ["proc-macro2"]
