+++ tasks/track_memory_allocations/Cargo.toml
@@ -18,7 +18,6 @@ doctest = false
 
 [dependencies]
 oxc_allocator = { workspace = true, features = ["track_allocations"] }
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
 oxc_formatter = { workspace = true }
 oxc_minifier = { workspace = true }
 oxc_parser = { workspace = true }
@@ -33,7 +32,3 @@ saphyr = { workspace = true }
 [features]
 # Sentinel flag that should only get enabled if we're running under `--all-features`
 is_all_features = []
-
-# `oxc_ast` is a feature-only dependency, to enable its `disable_old_builder` feature
-[package.metadata.cargo-shear]
-ignored = ["oxc_ast"]
