+++ tasks/transform_checker/Cargo.toml
@@ -14,7 +14,7 @@ doctest = false
 
 [dependencies]
 oxc_allocator = { workspace = true }
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
+oxc_ast = { workspace = true }
 oxc_ast_visit = { workspace = true }
 oxc_diagnostics = { workspace = true }
 oxc_semantic = { workspace = true }
