+++ tasks/website_linter/Cargo.toml
@@ -15,7 +15,6 @@ doctest = false
 [dependencies]
 bpaf = { workspace = true, features = ["docgen"] }
 itertools = { workspace = true }
-oxc_ast = { workspace = true, features = ["disable_old_builder"] }
 oxc_linter = { workspace = true, features = ["ruledocs"] }
 oxlint = { workspace = true, features = ["ruledocs"] }
 pico-args = { workspace = true }
@@ -27,5 +26,4 @@ website_common = { workspace = true }
 insta = { workspace = true }
 
 [package.metadata.cargo-shear]
-# `oxc_ast` is a feature-only dependency, to enable its `disable_old_builder` feature
-ignored = ["bpaf", "oxc_ast"]
+ignored = ["bpaf"]
